Rename MarchingCubes implementation class names and reformat.
